Closes #142

+++ backend/src/main/java/com/example/cms/university/UniversityService.java
@@ -217,7 +217,7 @@ public class UniversityService {
         }
 
         validateForDelete(university);
-        pageRepository.delete(university.getMainPage());
+        if (university.getMainPage() != null) pageRepository.delete(university.getMainPage());
         universityRepository.delete(university);
     }
